harness/pyproject.toml
[project]
name = "conversion-harness"
version = "0.1.0"
description = "Act I evaluation harness wrapping tau2-bench (traces, scores, Langfuse)."
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
  "typer>=0.12.0",
  "pyyaml>=6.0",
  "rich>=13.0",
  "pydantic>=2.0",
  "numpy>=1.24.0",
  "python-dotenv>=1.0.0",
  "langfuse>=4.5.0",
  "tau2",
]

[project.optional-dependencies]
dev = ["jsonschema>=4.0", "pytest>=8.0"]

[project.scripts]
harness = "harness.cli:app"

[tool.uv.sources]
tau2 = { path = "../tau2-bench", editable = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/harness"]
